backend/requirements.txt
asgiref==3.4.1
Django==3.2
django-debug-toolbar==3.2.1
django-cors-headers==3.13.0
djangorestframework==3.12.4
djoser==2.1.0
drf-extra-fields==3.4.0
django-filter==2.4.0
python-dotenv==0.21.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.3.1 
Pillow==8.3.1